Grading
Great app, but you didn't actually use an integrated API, not even to build your user input selections!

Still very well done, with great UI and visuals, you just missed the point of the prompt. There's essentially no difference between what you did in loading the data and having the dataset saved in your file. I would say that using the download url isn't even the API, as it's not a documented endpoint.

It was great having you in class. good luck with the rest of your semester!